+++ Demo/Demo/Extensions/PaymentButtonEnums+Extension.swift
@@ -36,7 +36,7 @@ extension PayPalCreditButton.Color: CaseIterable {
 extension PaymentButtonEdges: CaseIterable {
 
     public static var allCases: [PaymentButtonEdges] {
-        [.rectangle, .rounded, .rounded, .custom(10)]
+        [.hardEdges, .rounded, .rounded, .custom(10)]
     }
 
     static func allCasesAsString() -> [String] {