+++ Sources/PaymentButtons/PaymentButton.swift
@@ -37,7 +37,6 @@ public class PaymentButton: UIButton {
         self.insets = insets
         self.label = label
         super.init(frame: .zero)
-        configure()
     }
 
     // MARK: - Views
@@ -115,7 +114,7 @@ public class PaymentButton: UIButton {
         switch size {
 
         case .mini:
-            return 24.0
+            return 11.0
 
         case .standard:
             return 20.0
@@ -238,10 +237,21 @@ public class PaymentButton: UIButton {
     // MARK: - Override Function
     override public func layoutSubviews() {
         super.layoutSubviews()
+        configure()
         if size == .mini {
             containerView.layer.cornerRadius = 4.0
         } else {
             containerView.layer.cornerRadius = edges.cornerRadius(for: containerView)
+        }
+    }
+
+    public override var intrinsicContentSize: CGSize {
+        switch size {
+        case .mini:
+           return CGSize(width: 36, height: 24)
+
+        case .standard:
+            return CGSize(width: frame.size.width, height: imageHeight + size.elementPadding.top * 2)
         }
     }